cleanup & some more tests
