+++ crates/iceberg-catalog/src/api/management/mod.rs
@@ -317,7 +317,7 @@ pub mod v1 {
     }
 
     /// Type of tabular
-    #[derive(Debug, Serialize, Clone, Copy, utoipa::ToSchema, strum::Display)]
+    #[derive(Debug, Serialize, Clone, Copy, utoipa::ToSchema, strum::Display, PartialEq, Eq)]
     #[serde(rename_all = "kebab-case")]
     pub enum TabularType {
         Table,
